# config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "Imago Management Api",
        "description": "A REST Api to access digital twins."
    },
    "schemes": [
        "http",
        "https"
    ],
    "basePath": "/",
    "paths": {
...
        "/device/{deviceId}": {
            "put": {
                "description": "Registers a new device on the IoTHub Device registry",
                "tags": [
                    "Device"
                ],
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "operationId": "device_registerDevice",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "item": {
                                "$ref": "#/definitions/Device"
                            }
                        }
                    }
                },
                "deprecated": false
            },
            "get": {
                "description": "Fetches the device with the given id from the IoTHub registry",
                "tags": [
                    "Device"
                ],
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "operationId": "device_getDeviceById",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "item": {
                                "$ref": "#/definitions/Device"
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND"
                    }
                },
                "deprecated": false
            },
            "delete": {
                "description": "Deletes the device with the given id from the IoTHub registry",
                "tags": [
                    "Device"
                ],
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "operationId": "device_deleteDevice",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "Device": {
            "type": "object",
            "properties": {}
        },
...
    }
}